seperate: throwMonsters, throwItems, detectAttackCollision, killMonster
